done refactoring for now
