+++ lib/client/helper/apicommunication.rb
@@ -7,8 +7,9 @@ module Vermillion
 
         begin
           resp = @network.post(endpoint.to_s, server[:key])
-          # puts http + server[:address] + endpoint + args_qs
-          # puts resp.body
+
+          Notify.spit(endpoint.to_s) if DEBUG
+          Notify.spit(resp.body) if DEBUG
 
           # generic failure for invalid response type
           raise Errno::ECONNREFUSED if resp["Content-Type"] != "application/json"
@@ -40,9 +41,10 @@ module Vermillion
       def setup(input, endpoint, args = {})
         server_name = input
         remote_site = nil
-        server_name, remote_site, other = input.split('/') if input.to_s.include?('/')
+        server_name, remote_site = input.split('/') if input.to_s.include?('/')
 
         server = @config.get(:servers).select { |hash| hash[:name].to_sym == server_name.to_sym }.first
+
         # warn user if the server is not defined
         return Notify.warning("Server not found: #{server_name}") unless server
         # warn user if the site does not have a secret key property set
@@ -54,7 +56,6 @@ module Vermillion
         endpoint.server = server[:address]
         endpoint.protocol = server[:https]
         endpoint.add(:remote, remote_site) if remote_site
-        endpoint.add(:other, "/#{other}") if other
         endpoint.add(:query_string, Utils.to_query_string(args))
 
         [server, endpoint]
